+++ src/Common/ExecTrait.php
@@ -368,11 +368,26 @@ trait ExecTrait
     protected function printAction($context = [])
     {
         $command = $this->getCommandDescription();
+        $formatted_command = $this->formatCommandDisplay($command);
+
         $dir = $this->workingDirectory ? " in {dir}" : "";
         $this->printTaskInfo("Running {command}$dir", [
-                'command' => $command,
+                'command' => $formatted_command,
                 'dir' => $this->workingDirectory
             ] + $context);
+    }
+
+    /**
+     * @param $command
+     *
+     * @return mixed
+     */
+    protected function formatCommandDisplay($command)
+    {
+        $formatted_command = str_replace("&&", "&&\n", $command);
+        $formatted_command = str_replace("||", "||\n", $formatted_command);
+
+        return $formatted_command;
     }
 
     /**