+++ tests/unit/Task/ApiGenTest.php
@@ -47,7 +47,7 @@ class ApiGenTest extends \Codeception\TestCase\Test
 
         $winCmd = 'apigen generate --config ./apigen.neon --source src --extensions php --exclude test --exclude tmp --skip-doc-path a --skip-doc-path b --charset "utf8,iso88591" --internal no --php yes --tree yes --debug no';
 
-        $cmd = (strtoupper(substr(PHP_OS, 0, 3)) === 'WIN') ? $winCmd : $linuxCmd;
+        $cmd = stripos(PHP_OS, 'WIN') === 0 ? $winCmd : $linuxCmd;
 
         verify($task->getCommand())->equals($cmd);