+++ tests/unit/Task/PHPServerTest.php
@@ -41,18 +41,16 @@ class PHPServerTest extends \Codeception\TestCase\Test
 
     public function testServerCommand()
     {
-        if (strtolower(PHP_OS) === 'linux') {
-            $expectedCommand = 'exec php -S 127.0.0.1:8000 -t web';
-        } else {
-            $expectedCommand = 'php -S 127.0.0.1:8000 -t web';
-        }
+        // There is an 'exec ' at the beginning of the command here when
+        // running on Linux. Windows and MacOS do not have this prefix.
+        $cmd = stripos(PHP_OS, 'linux') === false ? '' : 'exec ';
+        $cmd .= 'php -S 127.0.0.1:8000 -t web';
 
         verify(
             (new \Robo\Task\Development\PhpServer('8000'))
                 ->host('127.0.0.1')
                 ->dir('web')
                 ->getCommand()
-        )->equals($expectedCommand);
+        )->equals($cmd);
     }
-
 }